hugeee push
so during my process i was really caught up in trying to make the code work out I forgot to push regularly so this is a huge update.

Despite how it seem, this game isn't interactive all all but it almost like the charcter interacts with the player. Using txt files, audio, video, and console tease had allowed me to make this creppy meta/4th wall type of narrtive which I really enjoy.

I was able to impletment txt string dialouge, save files, and sound to add to the atmostphere

I really struggled on trying to make a typewriter animation for the text. Sadly I never got it to work and spend hours on these effect for nothing :(

I did divate from orginal plans which I think turned out well.

In my process i utilzied txt files to make onscreen and off screen dialouge

Using the console to directly speak to the user if definelty my proud effects. Its so invasive yet distant.

in open processing at least one of my sound files crash the whole browers thing but also makes it disorted when it was supposed to be cute

this was a plus and a loss as i wanted to introduce creepy vibes later in my game.

Enjoy this huge push, and see my easter eggs in the open processing files :))